+++ sketch-core/src/jsCommonTest/kotlin/com/github/panpf/sketch/core/jscommon/test/cache/MemoryCacheJsCommonTest.kt
@@ -1,18 +1,17 @@
 package com.github.panpf.sketch.core.jscommon.test.cache
 
 import com.github.panpf.sketch.PlatformContext
-import com.github.panpf.sketch.cache.defaultMemoryCacheSizePercent
+import com.github.panpf.sketch.cache.defaultMemoryCacheSize
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
 class MemoryCacheJsCommonTest {
 
     @Test
-    fun testDefaultMemoryCacheSizePercent() {
+    fun testDefaultMemoryCacheSize() {
         assertEquals(
-            expected = 0.20,
-            actual = PlatformContext.INSTANCE.defaultMemoryCacheSizePercent(),
-            absoluteTolerance = 0.0
+            expected = 256L * 1024 * 1024,
+            actual = PlatformContext.INSTANCE.defaultMemoryCacheSize(),
         )
     }
 }